rename adapt_prompts and remove optimize_prompt
