+++ docs/docs/genai/getting-started/index.mdx
@@ -109,45 +109,49 @@ Register prompts and automatically optimize them with data-driven techniques:
 
 ```python
 import mlflow
-from mlflow.genai.optimize import OptimizerConfig, LLMParams
+import openai
+from mlflow.genai.optimize import LLMParams
 
 # Register an initial prompt
-initial_prompt = mlflow.genai.register_prompt(
+prompt = mlflow.genai.register_prompt(
     name="math_tutor",
     template="Answer this math question: {{question}}. Provide a clear explanation.",
 )
 
-# Prepare training data for optimization
+
+# Define prediction function
+def predict_fn(question: str) -> str:
+    prompt = mlflow.genai.load_prompt("prompts:/math_tutor/latest")
+    completion = openai.OpenAI().chat.completions.create(
+        model="gpt-4o-mini",
+        messages=[{"role": "user", "content": prompt.format(question=question)}],
+    )
+    return completion.choices[0].message.content
+
+
+# Prepare training data with inputs and expected outputs
 train_data = [
-    {"question": "What is 15 + 27?", "expected": "42"},
-    {"question": "Calculate 8 × 9", "expected": "72"},
+    {"inputs": {"question": "What is 15 + 27?"}, "outputs": "42"},
+    {"inputs": {"question": "Calculate 8 × 9"}, "outputs": "72"},
+    {"inputs": {"question": "What is 100 - 37?"}, "outputs": "63"},
     # ... more examples
 ]
 
-# Automatically optimize the prompt using MLflow + DSPy
-result = mlflow.genai.optimize_prompt(
-    target_llm_params=LLMParams(model_name="openai/gpt-4o-mini"),
-    prompt=initial_prompt,
+# Automatically optimize the prompt using MLflow + GEPA
+result = mlflow.genai.optimize_prompts(
+    predict_fn=predict_fn,
     train_data=train_data,
-    eval_data=train_data[:5],  # Hold-out evaluation set
-    optimizer_config=OptimizerConfig(
-        num_instruction_candidates=5,  # Try 5 different prompt variations
-        max_few_shot_examples=3,  # Include up to 3 examples
-    ),
+    target_prompt_uris=[prompt.uri],
+    optimizer_lm_params=LLMParams(model_name="openai:/gpt-4o-mini"),
 )
 
-# The optimized prompt is automatically registered as a new version of the original prompt
-optimized_prompt = result.optimized_prompt
-print(
-    f"Optimization improved accuracy from {result.baseline_score:.2f} to {result.optimized_score:.2f}"
-)
-print(
-    f"Optimized prompt registered as version {optimized_prompt.version} of '{optimized_prompt.name}'"
-)
+# The optimized prompt is automatically registered as a new version
+optimized_prompt = result.optimized_prompts[0]
+print(f"Optimized prompt registered as version {optimized_prompt.version}")
+print(f"Template: {optimized_prompt.template}")
 
-# Deploy the best-performing version
+# Deploy the optimized version
 with mlflow.start_run():
-    # Use the optimized prompt in your application
     model_info = mlflow.openai.log_model(
         model="gpt-4o-mini",
         task="llm/v1/completions",
@@ -164,8 +168,7 @@ Transform manual prompt engineering into systematic, data-driven optimization wi
 Ready to get started? You'll need:
 
 - Python 3.10+ installed
-- MLflow 3.0+ (`pip install --upgrade mlflow`)
-- For prompt optimization: DSPy (`pip install dspy`)
+- MLflow 3.5+ (`pip install --upgrade mlflow`)
 - API access to an LLM provider (OpenAI, Anthropic, etc.)
 
 ---
