+++ mlflow/genai/optimize/__init__.py
@@ -1,8 +1,5 @@
-from mlflow.genai.optimize.adapt import adapt_prompts
-from mlflow.genai.optimize.adapters import BasePromptAdapter
-from mlflow.genai.optimize.base import optimize_prompt
-from mlflow.genai.optimize.optimizers import BasePromptOptimizer, DSPyPromptOptimizer
-from mlflow.genai.optimize.optimizers.utils import format_dspy_prompt
+from mlflow.genai.optimize.adapt import optimize_prompts
+from mlflow.genai.optimize.optimizers import BasePromptOptimizer, GepaPromptOptimizer
 from mlflow.genai.optimize.types import (
     EvaluationResultRecord,
     LLMParams,
@@ -14,17 +11,14 @@ from mlflow.genai.optimize.types import (
 )
 
 __all__ = [
-    "optimize_prompt",
+    "optimize_prompts",
     "OptimizerConfig",
     "LLMParams",
     "PromptOptimizationResult",
     "OptimizerOutput",
     "BasePromptOptimizer",
-    "DSPyPromptOptimizer",
-    "format_dspy_prompt",
-    "adapt_prompts",
+    "GepaPromptOptimizer",
     "EvaluationResultRecord",
-    "BasePromptAdapter",
     "PromptAdapterOutput",
     "PromptAdaptationResult",
 ]
