--- mlflow/genai/optimize/optimizers/dspy_optimizer.py
@@ -1,206 +1,0 @@
-import importlib.metadata
-import importlib.util
-import logging
-from typing import TYPE_CHECKING, Any, Callable, Optional
-
-from packaging.version import Version
-
-from mlflow.entities.model_registry import PromptVersion
-from mlflow.exceptions import MlflowException
-from mlflow.genai.optimize.optimizers import BasePromptOptimizer
-from mlflow.genai.optimize.optimizers.utils import parse_model_name
-from mlflow.genai.optimize.types import LLMParams, ObjectiveFn, OptimizerConfig, OptimizerOutput
-from mlflow.genai.optimize.util import create_metric_from_scorers, infer_type_from_value
-from mlflow.genai.scorers import Scorer
-from mlflow.utils.annotations import experimental
-
-if TYPE_CHECKING:
-    import dspy
-    import pandas as pd
-
-_logger = logging.getLogger(__name__)
-
-
-@experimental(version="3.3.0")
-class DSPyPromptOptimizer(BasePromptOptimizer):
-    def __init__(self, optimizer_config: OptimizerConfig):
-        super().__init__(optimizer_config)
-
-        if not importlib.util.find_spec("dspy"):
-            raise ImportError("dspy is not installed. Please install it with `pip install dspy`.")
-
-        dspy_version = importlib.metadata.version("dspy")
-        if Version(dspy_version) < Version("2.6.0"):
-            raise MlflowException(
-                f"Current dspy version {dspy_version} is unsupported. "
-                "Please upgrade to version >= 2.6.0"
-            )
-
-    def optimize(
-        self,
-        prompt: PromptVersion,
-        target_llm_params: LLMParams,
-        train_data: "pd.DataFrame",
-        scorers: list[Scorer],
-        objective: ObjectiveFn | None = None,
-        eval_data: Optional["pd.DataFrame"] = None,
-    ) -> OptimizerOutput:
-        import dspy
-
-        _logger.info(
-            f"🎯 Starting prompt optimization for: {prompt.uri}\n"
-            f"⏱️ This may take several minutes or longer depending on dataset size...\n"
-            f"📊 Training with {len(train_data)} examples."
-        )
-
-        input_fields = self._get_input_fields(train_data)
-        self._validate_input_fields(input_fields, prompt)
-        output_fields = self._get_output_fields(train_data)
-
-        lm = dspy.LM(
-            model=parse_model_name(target_llm_params.model_name),
-            temperature=target_llm_params.temperature,
-            api_base=target_llm_params.base_uri,
-        )
-
-        if self.optimizer_config.optimizer_llm:
-            teacher_lm = dspy.LM(
-                model=parse_model_name(self.optimizer_config.optimizer_llm.model_name),
-                temperature=self.optimizer_config.optimizer_llm.temperature,
-                api_base=self.optimizer_config.optimizer_llm.base_uri,
-            )
-        else:
-            teacher_lm = lm
-
-        if self.optimizer_config.extract_instructions:
-            instructions = self._extract_instructions(prompt.template, teacher_lm)
-        else:
-            instructions = prompt.template
-
-        signature = dspy.make_signature(
-            {key: (type_, dspy.InputField()) for key, type_ in input_fields.items()}
-            | {key: (type_, dspy.OutputField()) for key, type_ in output_fields.items()},
-            instructions,
-        )
-
-        # Define main student program
-        program = dspy.Predict(signature)
-        adapter = dspy.JSONAdapter()
-
-        train_data = self._convert_to_dspy_dataset(train_data)
-        eval_data = self._convert_to_dspy_dataset(eval_data) if eval_data is not None else None
-
-        with dspy.context(lm=lm, adapter=adapter):
-            return self.run_optimization(
-                prompt=prompt,
-                program=program,
-                metric=self._convert_to_dspy_metric(
-                    input_fields, output_fields, scorers, objective
-                ),
-                train_data=train_data,
-                eval_data=eval_data,
-            )
-
-    def run_optimization(
-        self,
-        prompt: PromptVersion,
-        program: "dspy.Module",
-        metric: Callable[["dspy.Example", "dspy.Example", Any], float],
-        train_data: list["dspy.Example"],
-        eval_data: list["dspy.Example"],
-    ) -> OptimizerOutput:
-        """
-        Run the optimization process for the given prompt and program.
-
-        Parameters
-        ----------
-        prompt : PromptVersion
-            The prompt version to optimize.
-        program : dspy.Module
-            The DSPy program/module to optimize.
-        metric : Callable[[dspy.Example], float]
-            A callable that computes a metric score for a given example.
-        train_data : list[dspy.Example]
-            List of training examples for optimization.
-        eval_data : list[dspy.Example]
-            List of evaluation examples for validation.
-
-        Returns
-        -------
-        OptimizerOutput
-            The result of the optimization, including the optimized prompt and metrics.
-
-        Raises
-        ------
-        NotImplementedError
-            This method must be implemented by subclasses.
-        """
-        raise NotImplementedError(
-            "Subclasses of DSPyPromptOptimizer must implement `run_optimization`."
-        )
-
-    def _get_input_fields(self, train_data: "pd.DataFrame") -> dict[str, type]:
-        if "inputs" in train_data.columns:
-            sample_input = train_data["inputs"].values[0]
-            return {k: infer_type_from_value(v) for k, v in sample_input.items()}
-        return {}
-
-    def _get_output_fields(self, train_data: "pd.DataFrame") -> dict[str, type]:
-        if "expectations" in train_data.columns:
-            sample_output = train_data["expectations"].values[0]
-            return {k: infer_type_from_value(v) for k, v in sample_output.items()}
-        return {}
-
-    def _convert_to_dspy_dataset(self, data: "pd.DataFrame") -> list["dspy.Example"]:
-        import dspy
-
-        examples = []
-        for _, row in data.iterrows():
-            expectations = row["expectations"] if "expectations" in row else {}
-            examples.append(
-                dspy.Example(**row["inputs"], **expectations).with_inputs(*row["inputs"].keys())
-            )
-        return examples
-
-    def _convert_to_dspy_metric(
-        self,
-        input_fields: dict[str, type],
-        output_fields: dict[str, type],
-        scorers: list[Scorer],
-        objective: ObjectiveFn | None = None,
-    ) -> Callable[["dspy.Example", "dspy.Example", Any], float]:
-        # Create the base metric using the utility function
-        base_metric = create_metric_from_scorers(scorers, objective)
-
-        def metric(example: "dspy.Example", pred: "dspy.Example", trace=None) -> float:
-            inputs = {key: example.get(key) for key in input_fields.keys()}
-            expectations = {key: example.get(key) for key in output_fields.keys()}
-            outputs = {key: pred.get(key) for key in output_fields.keys()}
-
-            return base_metric(inputs, outputs, expectations)
-
-        return metric
-
-    def _validate_input_fields(self, input_fields: dict[str, type], prompt: PromptVersion) -> None:
-        if missing_fields := set(prompt.variables) - set(input_fields.keys()):
-            raise MlflowException(
-                f"Validation failed. Missing prompt variables in dataset: {missing_fields}. "
-                "Please ensure your dataset contains columns for all prompt variables."
-            )
-
-    def _extract_instructions(self, template: str | dict[str, Any], lm: "dspy.LM") -> str:
-        import dspy
-
-        extractor = dspy.Predict(
-            dspy.make_signature(
-                {
-                    "prompt": (str, dspy.InputField()),
-                    "instruction": (str, dspy.OutputField()),
-                },
-                "Extract the core instructions from the prompt "
-                "to use as the system message for the LLM.",
-            )
-        )
-
-        with dspy.context(lm=lm):
-            return extractor(prompt=template).instruction
