+++ mlflow/genai/optimize/optimizers/utils/__init__.py
@@ -1,6 +1,15 @@
-from mlflow.genai.optimize.optimizers.utils.dspy_optimizer_utils import (
-    format_dspy_prompt,
-    parse_model_name,
-)
+def parse_model_name(model_name: str) -> str:
+    """
+    Parse model name from MLflow format to provider/model format.
 
-__all__ = ["format_dspy_prompt", "parse_model_name"]
+    Converts "provider:/model" to "provider/model" format.
+
+    Args:
+        model_name: Model name in either "provider:/model" or "provider/model" format
+
+    Returns:
+        Model name in "provider/model" format
+    """
+    if ":/" in model_name:
+        return model_name.replace(":/", "/")
+    return model_name
